backend/requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.43.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kaleido==0.2.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.26.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
plotly==5.18.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
scipy==1.11.3
six==1.16.0
tenacity==8.2.3
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
